backend/requirements.txt
# Requirements for Test Impact Analysis System
# Python 3.7+ required

# PostgreSQL database adapter
# psycopg2-binary is recommended for easier installation (includes all dependencies)
# For production, you might want to use psycopg2 instead
psycopg2-binary>=2.9.0

# Environment variable management (.env file support)
python-dotenv>=1.0.0

# Web Framework (FastAPI)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # Required for file uploads (UploadFile, Form)

# Enhanced AST parsing (for better docstring extraction)
# Uncomment if needed for advanced parsing features
# ast-comments>=1.1.3

# Note: Most functionality uses Python standard library:
# - ast (Abstract Syntax Tree parsing)
# - pathlib (File path handling)
# - json (JSON file operations)
# - re (Regular expressions)
# - datetime (Timestamp generation)
# - collections (Data structures)
# - typing (Type hints)
# - contextlib (Context managers)
# - sys, os (System operations)
# - zipfile (ZIP file handling)
# - hashlib (SHA-256 hashing)
# - shutil (File operations)

# Vector database for semantic search
# Pinecone is the only supported backend
pinecone>=3.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.24.0

# Tree-sitter for multi-language parsing 
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
tree-sitter-javascript>=0.20.0
tree-sitter-java>=0.20.0
tree-sitter-typescript>=0.20.0
# C / C++ (optional; pip may upgrade tree-sitter — if install fails, omit these)
tree-sitter-c>=0.21.0
tree-sitter-cpp>=0.23.0
